% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdi_funcs.R
\name{comp_G_hat}
\alias{comp_G_hat}
\title{calculating inverse weights using the KM method}
\usage{
comp_G_hat(dat, tau)
}
\arguments{
\item{tau}{}
}
\value{

}
\description{
calculating inverse weights using the KM method
}
