% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdi_funcs.R
\name{comp_scores}
\alias{comp_scores}
\title{fit Fine-Gray model on mdat and use the fitted model to generate prediction for pdat}
\usage{
comp_scores(mdat, pdat, zcov, tau, n_l)
}
\arguments{
\item{n_l}{}
}
\value{

}
\description{
fit Fine-Gray model on mdat and use the fitted model to generate prediction for pdat
}
