% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdi_funcs.R
\name{fine_gray_pdi}
\alias{fine_gray_pdi}
\title{computing PDI with competing risks data using Fine-Gray model}
\usage{
fine_gray_pdi(dat, tau, n_l, seed, cv_rep, zcov)
}
\arguments{
\item{dat}{data}

\item{tau}{the prespcified inspection time}

\item{n_l}{number of outcome categories, including event-free}

\item{seed}{random seed for cross validation}

\item{cv_rep}{number of cross validation repetitions}

\item{zcov}{covariates used in the FG model}
}
\value{

}
\description{
computing PDI with competing risks data using Fine-Gray model
}
