% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdi_funcs.R
\name{pdi}
\alias{pdi}
\title{pdi function}
\usage{
pdi(tau, scores, x, delt_eps, G_hat, n_l)
}
\arguments{
\item{scores}{a n by n_l matrix of prognostic scores}

\item{x}{a vector of observation times}

\item{delt_eps}{a vector of event indicators}

\item{G_hat}{a vector of estimated inverse weights}

\item{n_l}{}
}
\value{

}
\description{
pdi function
}
